+++ tests/testthat/test_Measure.R
@@ -128,5 +128,5 @@ test_that("time_train is > 0", {
   skip_on_cran()
   rr = resample(tsk("iris"), lrn("classif.debug"), rsmp("holdout"))
   res = rr$score(msr("time_train"))
-  expect_gt(res$time_train, 0)
+  expect_gte(res$time_train, 0)
 })